+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Server metadata supports `hasStaticData` and `maxRecordCount`
+
 ## [2.2.0] - 2017-06-05
 ### Added
 * Support displayField, timeInfo, maxRecordCount, idField